Split the code and add unit tests
