+++ Cargo.toml
@@ -14,3 +14,7 @@ clap = { version = "4.0", features = ["derive"] }
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 mime_guess = "2.0"
+
+[dev-dependencies]
+hyper = { version = "0.14", features = ["full"] }
+tower = { version = "0.5.2", features = ["make"] }
